ART - MASSIVE Enviro Overhaul + Asset Store Assets
I made a Terrain for some nice height differences in the scene.
Next, I started downloading a lot of EULA-compliant assets from the Asset Store as placeholders for some of the elements in the environment.
Then, I hooked a few SelectableObjects in the scene, and prettied them up with some particle work. I made some firefly particles as well that I want to use in the night.

I'm not done yet, I have some modelling I want to get done quickly, then some writing with the SelectableObject data in the scene.

:llama: